% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdverseEvents.R
\name{AE}
\alias{AE}
\title{Define an Adverse Event}
\usage{
AE(
  time_to_first_event,
  event_duration,
  time_between_events,
  event_prob = 1,
  severity_probability = SEVERITY_PROBABILITIES$equal
)
}
\arguments{
\item{time_to_first_event}{Expected time until the first event (in days)}

\item{event_duration}{Expected duration of each event (in days)}

\item{time_between_events}{Expected time between events (in days )}

\item{event_prob}{Probability of an event happening at all (0 to 1)}

\item{severity_probability}{Probability distribution for event severity (vector of probabilities)}
}
\value{
An event object with defined parameters
}
\description{
This function defines an event with parameters such as time to the first event,
event duration, time between events, event probability, and severity probabilities.
}
