% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdverseEvents.R
\name{simulate_event}
\alias{simulate_event}
\title{Simulate a single event}
\usage{
simulate_event(event, max_time)
}
\arguments{
\item{event}{An event object}

\item{max_time}{Maximum time to simulate events (in days)}
}
\value{
An event object with simulated event times
}
\description{
This function simulates an event based on the provided event parameters.
}
\seealso{
\code{\link{AE}} for defining an event
}
